+++ HISTORY.md
@@ -3,6 +3,7 @@
 ------------
 
  * Fixed; Replaced helpers.jsonRequest() with SuperAgent due to a bug in JQuery $.ajax().
+ * Added; Sage and bump functions for posts.
 
 0.0.2 / 2012-09-10
 ------------------